+++ solys2tracker/common.py
@@ -12,10 +12,11 @@ import random
 import string
 
 """___Third-Party Modules___"""
-from PySide2 import QtWidgets, QtCore
+from PySide2 import QtWidgets, QtCore, QtGui
 from matplotlib.axes import Axes
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.figure import Figure
+from asdcontroller import asd_types as asdt
 
 """___Solys2Tracker Modules___"""
 # import here
@@ -113,6 +114,9 @@ class MplCanvas(FigureCanvas):
 class GraphWidget(QtWidgets.QWidget):
     def __init__(self):
         super().__init__()
+        self.title = ""
+        self.xlabel = ""
+        self.ylabel = ""
         self._build_layout()
 
     def _build_layout(self):
@@ -123,14 +127,92 @@ class GraphWidget(QtWidgets.QWidget):
     def update_plot(self, x_data: list, y_data: list):
         self.x_data = x_data
         self.y_data = y_data
-        self.canvas.axes.cla()  # Clear the canvas.
-        self.canvas.axes.plot(self.x_data, self.y_data, "r")
-        self.canvas.draw()
+        self.redraw()
 
     def update_labels(self, title: str, xlabel: str, ylabel: str):
-        self.canvas.axes.set_title(title)
-        self.canvas.axes.set_xlabel(xlabel)
-        self.canvas.axes.set_ylabel(ylabel)
+        self.title = title
+        self.xlabel = xlabel
+        self.ylabel = ylabel
+        self.redraw()
+    
+    def redraw(self):
+        self.canvas.axes.cla()  # Clear the canvas.
+        self.canvas.axes.plot(self.x_data, self.y_data, "o",  markersize=2)
+        self.canvas.axes.set_title(self.title)
+        self.canvas.axes.set_xlabel(self.xlabel)
+        self.canvas.axes.set_ylabel(self.ylabel)
+        self.canvas.draw()
+
+class CaptureDataWindow(QtWidgets.QWidget):
+    def __init__(self):
+        super().__init__()
+        self._build_layout()
+    
+    def _build_layout(self):
+        self.v_spacers = []
+        self.h_spacers = []
+        self.main_layout = QtWidgets.QVBoxLayout(self)
+        # Headers
+        self.headers_layout = QtWidgets.QFormLayout()
+        self.label_itime = QtWidgets.QLabel("Integration time:")
+        self.label_drift = QtWidgets.QLabel("VNIR Drift:")
+        self.label_gain1 = QtWidgets.QLabel("Swir1 gain:")
+        self.label_gain2 = QtWidgets.QLabel("Swir2 gain:")
+        self.label_v_itime = QtWidgets.QLabel("")
+        self.label_v_drift = QtWidgets.QLabel("")
+        self.label_v_gain1 = QtWidgets.QLabel("")
+        self.label_v_gain2 = QtWidgets.QLabel("")
+        self.headers_layout.addRow(self.label_itime, self.label_v_itime)
+        self.headers_layout.addRow(self.label_drift, self.label_v_drift)
+        self.headers_layout.addRow(self.label_gain1, self.label_v_gain1)
+        self.headers_layout.addRow(self.label_gain2, self.label_v_gain2)
+        # Graph
+        self.graph = GraphWidget()
+        # Finish main layout
+        add_spacer(self.main_layout, self.v_spacers)
+        self.main_layout.addLayout(self.headers_layout)
+        add_spacer(self.main_layout, self.v_spacers)
+        self.main_layout.addWidget(self.graph)
+        add_spacer(self.main_layout, self.v_spacers)
+
+    def update_plot(self, x_data: list, y_data: list):
+        self.graph.update_plot(x_data, y_data)
+
+    def update_labels(self, title: str, xlabel: str, ylabel: str):
+        self.graph.update_labels(title, xlabel, ylabel)
+    
+    def update_headers(self, spec: asdt.FRInterpSpec):
+        vh = spec.fr_spectrum_header.v_header
+        s1 = spec.fr_spectrum_header.s1_header
+        s2 = spec.fr_spectrum_header.s2_header
+        self.label_v_itime.setText(asdt.ITimeEnum(vh.it).to_str())
+        self.label_v_drift.setText(str(vh.drift))
+        self.label_v_gain1.setText(str(s1.gain))
+        self.label_v_gain2.setText(str(s2.gain))
+
+class GraphWindow(QtWidgets.QMainWindow):
+    def __init__(self):
+        super().__init__()
+        self.graph = CaptureDataWindow()
+        self.should_close = False
+        self.setCentralWidget(self.graph)
+    
+    def closeEvent(self, event: QtGui.QCloseEvent):
+        if self.should_close:
+            super().closeEvent(event)
+    
+    def force_close(self):
+        self.should_close = True
+        self.close()
+
+    def update_plot(self, x_data: list, y_data: list):
+        self.graph.update_plot(x_data, y_data)
+
+    def update_labels(self, title: str, xlabel: str, ylabel: str):
+        self.graph.update_labels(title, xlabel, ylabel)
+    
+    def update_headers(self, spec: asdt.FRInterpSpec):
+        self.graph.update_headers(spec)
 
 def _gen_random_str(len: int) -> str:
     """